ajustando deletes

+++ src/main/java/com/provodromo/provodromo/service/TipoUsuarioService.java
@@ -1,12 +1,15 @@
 package com.provodromo.provodromo.service;
 
 import com.provodromo.provodromo.model.TipoUsuario;
+import com.provodromo.provodromo.model.Usuario;
 import com.provodromo.provodromo.repository.TipoUsuarioRepository;
+import com.provodromo.provodromo.repository.UsuarioRepository;
 import com.provodromo.provodromo.service.base.BaseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 @Service
@@ -14,6 +17,9 @@ public class TipoUsuarioService implements BaseService<TipoUsuario> {
 
     @Autowired
     private TipoUsuarioRepository repository;
+
+    @Autowired
+    private UsuarioRepository usuarioRepository;
 
     @Override
     public TipoUsuario findById(Long id) {
@@ -34,7 +40,20 @@ public class TipoUsuarioService implements BaseService<TipoUsuario> {
 
     @Override
     public void deleteById(Long id) {
-        repository.deleteById(id);
+        TipoUsuario tipoUsuario = repository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Tipo de usuário não encontrado com o id: " + id));
+
+        if(!tipoUsuario.getNome().equals("Comum")) {
+            List<Usuario> usuarios = usuarioRepository.findByTipoUsuarioId(id);
+            TipoUsuario tipoComum = repository.findByName("Comum");
+
+            for (Usuario usuario : usuarios) {
+                usuario.setTipoUsuario(tipoComum);
+                usuarioRepository.save(usuario);
+            }
+
+            repository.deleteById(id);
+        }
     }
 
     public TipoUsuario findByName(String name) { return repository.findByName(name);}
